Refactor view layer into MVVM structure
