--- SearchBarView.swift
@@ -1,44 +1,0 @@
-import SwiftUI
-
-struct SearchBarView: View {
-    @State private var searchText: String = ""
-
-    var body: some View {
-        HStack(spacing: 16) {
-            Image(systemName: "magnifyingglass")
-                .foregroundStyle(.secondary)
-
-            TextField("Search", text: $searchText)
-                .textFieldStyle(.plain)
-                .textInputAutocapitalization(.none)
-                .disableAutocorrection(true)
-
-            Spacer(minLength: 0)
-
-            Image(systemName: "square.and.pencil")
-                .foregroundStyle(.blue)
-
-            Image(systemName: "notification")
-                .foregroundStyle(.orange)
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 12)
-        .background(
-            RoundedRectangle(cornerRadius: 20, style: .continuous)
-                .fill(Color(.systemGray6))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 20, style: .continuous)
-                .stroke(Color(.systemGray4))
-        )
-        .padding(.horizontal)
-    }
-}
-
-struct SearchBarView_Previews: PreviewProvider {
-    static var previews: some View {
-        SearchBarView()
-            .padding()
-            .previewLayout(.sizeThatFits)
-    }
-}
