--- View/PostCell.swift
@@ -1,1 +1,0 @@
-
