+++ View/Views/Components/Story/StorySectionView.swift
@@ -1,12 +1,5 @@
 import SwiftUI
 
-struct Story: Identifiable, Hashable {
-    let id = UUID()
-    let name: String
-    let avatarURL: URL
-    let thumbnailURL: URL
-}
-
 struct StorySectionView: View {
     let stories: [Story]
     var onStoryTap: (Story) -> Void = { _ in }
@@ -31,7 +24,7 @@ struct StorySectionView: View {
     }
 }
 
-struct StoryCell: View {
+private struct StoryCell: View {
     let story: Story
     var onTap: () -> Void
 
@@ -89,37 +82,5 @@ struct StoryCell: View {
             }
         }
         .buttonStyle(.plain)
-    }
-}
-
-struct StorySectionView_Previews: PreviewProvider {
-    static let demoStories: [Story] = [
-        Story(
-            name: "Zalo Official",
-            avatarURL: URL(string: "https://images.unsplash.com/photo-1544723795-3fb6469f5b39?auto=format&fit=crop&w=120&q=80")!,
-            thumbnailURL: URL(string: "https://images.unsplash.com/photo-1545239351-1141bd82e8a6?auto=format&fit=crop&w=400&q=80")!
-        ),
-        Story(
-            name: "SwiftUI Lover",
-            avatarURL: URL(string: "https://images.unsplash.com/photo-1521572163474-6864f9cf17ab?auto=format&fit=crop&w=120&q=80")!,
-            thumbnailURL: URL(string: "https://images.unsplash.com/photo-1504384308090-c894fdcc538d?auto=format&fit=crop&w=400&q=80")!
-        ),
-        Story(
-            name: "iOS Dev",
-            avatarURL: URL(string: "https://images.unsplash.com/photo-1500648767791-00dcc994a43e?auto=format&fit=crop&w=120&q=80")!,
-            thumbnailURL: URL(string: "https://images.unsplash.com/photo-1498050108023-c5249f4df085?auto=format&fit=crop&w=400&q=80")!
-        ),
-        Story(
-            name: "Design Team",
-            avatarURL: URL(string: "https://images.unsplash.com/photo-1527980965255-d3b416303d12?auto=format&fit=crop&w=120&q=80")!,
-            thumbnailURL: URL(string: "https://images.unsplash.com/photo-1521737604893-d14cc237f11d?auto=format&fit=crop&w=400&q=80")!
-        )
-    ]
-
-    static var previews: some View {
-        StorySectionView(stories: demoStories)
-            .padding(.vertical)
-            .previewLayout(.sizeThatFits)
-            .background(Color(.systemBackground))
     }
 }
