+++ View/Views/Diary/DiaryTabView.swift
@@ -1,0 +1,35 @@
+import SwiftUI
+
+struct DiaryTabView: View {
+    @ObservedObject var viewModel: DiaryTabViewModel
+
+    init(viewModel: DiaryTabViewModel = DiaryTabViewModel()) {
+        self.viewModel = viewModel
+    }
+
+    var body: some View {
+        ScrollView {
+            LazyVStack(spacing: 24) {
+                CreateMyStoryView(
+                    avatarURL: viewModel.currentUserAvatarURL,
+                    placeholder: viewModel.storyPlaceholder
+                )
+
+                StorySectionView(stories: viewModel.featuredStories)
+
+                ForEach(viewModel.activities) { activity in
+                    PostCellView(activity: activity)
+                        .padding(.horizontal)
+                }
+
+                AdsPostView(post: viewModel.adsPost) {}
+                    .padding(.horizontal)
+
+                SuggestStoryView(stories: viewModel.suggestedStories)
+                    .padding(.horizontal)
+            }
+            .padding(.vertical, 16)
+        }
+        .background(Color(.systemGroupedBackground))
+    }
+}
